+++ R/sites.R
@@ -1,4 +1,4 @@
-#' Descriptive Site Information
+#' Site Information
 #'
 #' @description Information for sites in the U.S. Geological Survey (USGS) monitoring networks,
 #'   Idaho National Laboratory (INL) and vicinity, Idaho.
@@ -29,12 +29,7 @@
 #'       "N" is not surveyed;
 #'       "S" is a transit, theodolite, or other surveying method; and
 #'       "X" is not applicable.}
-#'     \item{`coord_acy_cd`}{Accuracy code for latitude/longitude values.
-#'       The codes and their meanings are as follows:
-#'       "1" is accurate to \eqn{\pm}{+/-} 1 second;
-#'       "5" is accurate to \eqn{\pm}{+/-} 0.5 second;
-#'       "H" is accurate to \eqn{\pm}{+/-} 0.1 second; and
-#'       "S" is accurate to \eqn{\pm}{+/-} 1 second.}
+#'     \item{`coord_acy_va`}{Accuracy of latitude/longitude value in seconds.}
 #'     \item{`alt_va`}{Elevation of the land surface reference point,
 #'       in feet above the North American Vertical Datum of 1988 (NAVD 88).}
 #'     \item{`alt_meth_cd`}{Method code for measuring elevation.
@@ -105,9 +100,14 @@
 #'     \item{`pos`}{Position specifier for site-labels on a map.
 #'       Values of 1, 2, 3 and 4, respectively indicate positions below (south), to the left (west) of,
 #'       above (north), and to the right (east) of the site coordinates.}
+#'     \item{`min_dt`}{Date of first record.}
+#'     \item{`max_dt`}{Date of last record.}
+#'     \item{`nmeas`}{Number of measurments recorded at the site
+#'       includes either groundwater-level or streamflow measurements.}
+#'     \item{`nsamples`}{Number of water-quality samples collected from the site.}
+#'     \item{`nreps`}{Number of replicate samples collected from the site.}
 #'     \item{`geometry`}{Zero-dimensional geometry containing a single point.}
 #'   }
-#'   See [`crs`] dataset for coordinate reference system information.
 #'
 #' @source USGS site data acquired from the NWIS (U.S. Geological Survey, 2023).
 #'   The station name (`station_nm`) for the data was parsed for a more readable name.